Fix unnecessary null check
There's a bit of a broken logic here, what's the point of initialising colour if VibrantSwatch is null if we are just going to not use it if it's null anyway? Besides, we already null-checked before so why bother null-checking it here again?

This fixes the issue of detailed album view not having any background or text in it.