Remove redundant  API checks
We have a lot of these leftover codes that checks for API number. There's several problem to these API checks, one of them being we already satisfy these by default so we don't need checking.

Our min API is 21, which means we are already well above 16 by default, and the check for >= API 21 is also satisfied by this since there's no possibility that a device with API number lesser than 21 will use it.

Clean up the code and remove these because they seems redundant.