Better default artwork image
Our current default artwork image looks blurry and bad on higher resolution device as a result of it being on a very low resolution (300x300) by itself.

Let's create a new default arwork image on a higher resolution while not sacrificing much of the old artwork's style (also, change the tone of the artwork background to make it fit to our current primary theme colour).