Fix seconds duration display
There's a stupid bug caused by a derp in the code, because the "0" is displayed on the end, anything with seconds length lesser than 1 will display stupidly ("4:90" for example)
This is absolutely ridiculous because there's no way we'll ever get to >60  in seconds. Fix this by putting the "0" in front of the second in case the seconds length is lesser than 0.