[HAX] Fix repeatButton state when activity is first created
During the creation of the activity, isLooping is not defined yet, making the repeatButton stateless. The button then requires to be double tapped, hence toggling between the state of mediaPlayer.isLooping and setting the isLooping boolean from null to false.

To fix this, create a temporary hax that initialise isLooping boolean to false on init(), so when repeatButton is pressed, it properly toggles isLooping to true instead of from null to false.