Properly stop music if it reaches end duration
A small hack to stop music if end duration is reached, we make use of the seekBar and detect if it has reached the end duration, then we call intent to stop the music. This essentially stops the music and changes the "pause" drawable to "play".

There's a small issue with this implementation though, it's possible for the user to seek the seekBar further and the drawable won't change. I'll be addressing this in later commit.