Use default colours when vibrantSwatch is null
ColorGridTask has a problem where the text will be coloured *very* randomly even if the background colour is defaulted to card_background. For example, if the album doesn't have an album picture, the text colour (both albumName and albumDesc) will be coloured to black, this conflicts with the colour chosen for card_background (which is dark).

Fix this by having the colours set to lighter colour in case the vibrantSwatch returns null. There's still some problem here and there, but at least this fixes one of them.